server/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
google-generativeai==0.7.2
pandas==2.2.2
